backend/requirements.txt
django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
sqlparse==0.4.2
psycopg2-binary==2.9.3
pytz==2023.3
celery==5.3.0
redis==4.5.5